перенес BottomNavigationBar.kt в отдельный файл убрал ненужные Modifier удалил activity_main.xml
