+++ app/src/main/java/com/finance_tracker/finance_tracker/MainActivity.kt
@@ -6,20 +6,13 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
+import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
-import androidx.navigation.NavController
-import androidx.navigation.compose.currentBackStackEntryAsState
 import com.finance_tracker.finance_tracker.sreens.container.Container
-import com.finance_tracker.finance_tracker.theme.CoinTheme
 import com.finance_tracker.finance_tracker.theme.AppTheme
+import com.finance_tracker.finance_tracker.theme.CoinTheme
 
 class MainActivity : ComponentActivity() {
 
@@ -48,54 +41,10 @@ class MainActivity : ComponentActivity() {
         val permission = Manifest.permission.RECEIVE_SMS
         val grant = ContextCompat.checkSelfPermission(this, permission)
         if (grant != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, arrayOf(permission), REQUEST_CODE_SMS_PERMISSION)
-        }
-    }
-}
-
-@Composable
-fun BottomNavigationBar(
-    navController: NavController,
-    modifier: Modifier = Modifier
-) {
-    val items = remember {
-        listOf(
-            NavigationItem.Main,
-            NavigationItem.Operations,
-            NavigationItem.More
-        )
-    }
-    BottomNavigation(
-        modifier = modifier,
-        backgroundColor = CoinTheme.color.background,
-        contentColor = CoinTheme.color.content
-    ) {
-        val navBackStackEntry by navController.currentBackStackEntryAsState()
-        val currentRoute = navBackStackEntry?.destination?.route
-        items.forEach { item ->
-            BottomNavigationItem(
-                icon = {
-                    Icon(
-                        painterResource(id = item.iconRes),
-                        contentDescription = stringResource(id = item.title)
-                    )
-                },
-                label = { Text(text = stringResource(id = item.title)) },
-                selectedContentColor = CoinTheme.color.selectedContentColor,
-                unselectedContentColor = CoinTheme.color.unSelectedContentColor.copy(alpha = 0.4f),
-                alwaysShowLabel = true,
-                selected = currentRoute == item.route,
-                onClick = {
-                    navController.navigate(item.route) {
-                        navController.graph.startDestinationRoute?.let { route ->
-                            popUpTo(route) {
-                                saveState = true
-                            }
-                        }
-                        launchSingleTop = true
-                        restoreState = true
-                    }
-                }
+            ActivityCompat.requestPermissions(
+                this,
+                arrayOf(permission),
+                REQUEST_CODE_SMS_PERMISSION
             )
         }
     }
