+++ app/src/main/java/com/finance_tracker/finance_tracker/sreens/container/BottomNavigationBar.kt
@@ -1,0 +1,64 @@
+package com.finance_tracker.finance_tracker.sreens.container
+
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.navigation.NavController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import com.finance_tracker.finance_tracker.NavigationItem
+import com.finance_tracker.finance_tracker.theme.CoinTheme
+
+@Composable
+fun BottomNavigationBar(
+    navController: NavController,
+    modifier: Modifier = Modifier
+) {
+    val items = remember {
+        listOf(
+            NavigationItem.Main,
+            NavigationItem.Operations,
+            NavigationItem.More
+        )
+    }
+    BottomNavigation(
+        modifier = modifier,
+        backgroundColor = CoinTheme.color.background,
+        contentColor = CoinTheme.color.content
+    ) {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentRoute = navBackStackEntry?.destination?.route
+        items.forEach { item ->
+            BottomNavigationItem(
+                icon = {
+                    Icon(
+                        painterResource(id = item.iconRes),
+                        contentDescription = stringResource(id = item.title)
+                    )
+                },
+                label = { Text(text = stringResource(id = item.title)) },
+                selectedContentColor = CoinTheme.color.selectedContentColor,
+                unselectedContentColor = CoinTheme.color.unSelectedContentColor.copy(alpha = 0.4f),
+                alwaysShowLabel = true,
+                selected = currentRoute == item.route,
+                onClick = {
+                    navController.navigate(item.route) {
+                        navController.graph.startDestinationRoute?.let { route ->
+                            popUpTo(route) {
+                                saveState = true
+                            }
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                }
+            )
+        }
+    }
+}
